Move some ui components around
