+++ ui/src/components/auth/LoginForm.svelte
@@ -3,7 +3,7 @@
   import { user } from '../../stores';
   import LL from '../../i18n/i18n-svelte';
   import SolidButton from '../global/SolidButton.svelte';
-  import TextInput from '../global/TextInput.svelte';
+  import TextInput from '../forms/TextInput.svelte';
   import LockIcon from '../icons/LockIcon.svelte';
 
   export let registerLink = '';
