+++ ui/src/components/forms/LocaleInput.svelte
