+++ ui/src/components/logos/DomeLogo.svelte
