+++ ui/src/components/logos/DomeLogoLight.svelte
