+++ ui/src/components/logos/FullLogo.svelte
