+++ ui/src/components/logos/FullLogoLight.svelte
