+++ ui/src/components/poker/BecomeFacilitator.svelte
@@ -2,7 +2,7 @@
   import Modal from '../global/Modal.svelte';
   import LL from '../../i18n/i18n-svelte';
   import SolidButton from '../global/SolidButton.svelte';
-  import TextInput from '../global/TextInput.svelte';
+  import TextInput from '../forms/TextInput.svelte';
 
   export let toggleBecomeLeader = () => {};
   export let handleBecomeLeader = () => {};
