+++ ui/src/components/poker/VotingControls.svelte
@@ -2,7 +2,7 @@
   import SolidButton from '../global/SolidButton.svelte';
   import LL from '../../i18n/i18n-svelte';
   import { AppConfig } from '../../config';
-  import SelectInput from '../global/SelectInput.svelte';
+  import SelectInput from '../forms/SelectInput.svelte';
 
   export let sendSocketEvent = () => {};
   export let eventTag;
