+++ ui/src/components/storyboard/PersonasForm.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import SolidButton from '../global/SolidButton.svelte';
   import Modal from '../global/Modal.svelte';
-  import TextInput from '../global/TextInput.svelte';
+  import TextInput from '../forms/TextInput.svelte';
 
   export let toggleEditPersona = () => () => {};
   export let handlePersonaAdd = () => {};
