+++ ui/src/components/team/CreateTeam.svelte
@@ -2,7 +2,7 @@
   import Modal from '../global/Modal.svelte';
   import SolidButton from '../global/SolidButton.svelte';
   import LL from '../../i18n/i18n-svelte';
-  import TextInput from '../global/TextInput.svelte';
+  import TextInput from '../forms/TextInput.svelte';
 
   export let teamName = '';
