+++ ui/src/components/user/ProfileForm.svelte
@@ -10,11 +10,11 @@
   import SetupMFA from './SetupMFA.svelte';
   import DeleteConfirmation from '../global/DeleteConfirmation.svelte';
   import { user } from '../../stores';
-  import LocaleSwitcher from '../global/LocaleInput.svelte';
+  import LocaleSwitcher from '../forms/LocaleInput.svelte';
   import type { Locales } from '../../i18n/i18n-types';
   import { loadLocaleAsync } from '../../i18n/i18n-util.async';
-  import TextInput from '../global/TextInput.svelte';
-  import SelectInput from '../global/SelectInput.svelte';
+  import TextInput from '../forms/TextInput.svelte';
+  import SelectInput from '../forms/SelectInput.svelte';
 
   const setupI18n = async (locale: Locales) => {
     await loadLocaleAsync(locale);
