+++ ui/src/components/user/SetupMFA.svelte
@@ -3,7 +3,7 @@
   import SolidButton from '../global/SolidButton.svelte';
 
   import LL from '../../i18n/i18n-svelte';
-  import TextInput from '../global/TextInput.svelte';
+  import TextInput from '../forms/TextInput.svelte';
 
   export let toggleSetup = () => {};
   export let handleComplete = () => {};
