+++ ui/src/pages/retro/Retro.svelte
@@ -24,7 +24,7 @@
   import InviteUser from '../../components/retro/InviteUser.svelte';
   import PageLayout from '../../components/PageLayout.svelte';
   import UserAvatar from '../../components/user/UserAvatar.svelte';
-  import TextInput from '../../components/global/TextInput.svelte';
+  import TextInput from '../../components/forms/TextInput.svelte';
 
   export let retroId;
   export let notifications;
